% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{aggregate_matrix}
\alias{aggregate_matrix}
\title{Title}
\usage{
aggregate_matrix(input_matrix, rows = NULL, cols = NULL)
}
\arguments{
\item{input_matrix}{}

\item{rows}{}

\item{cols}{}
}
\description{
Title
}
