% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{append_googlesheet}
\alias{append_googlesheet}
\title{Title}
\usage{
append_googlesheet(
  new_data,
  ss,
  sheet,
  all_character = FALSE,
  ignore_cols = NULL
)
}
\arguments{
\item{new_data}{}

\item{ss}{}

\item{sheet}{}

\item{all_character}{}

\item{ignore_cols}{}
}
\description{
Title
}
