% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{apply_filters}
\alias{apply_filters}
\title{Title}
\usage{
apply_filters(
  input,
  filters = list(column = c("value1", "value2")),
  verbose = FALSE
)
}
\arguments{
\item{input}{}

\item{filters}{}
}
\description{
Title
}
