% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQUser.R
\name{createQUser}
\alias{createQUser}
\title{Create a specific Qualtrics user}
\usage{
createQUser(
  username,
  firstName,
  lastName,
  userType = "UT_6D93CHF3IYryJjn",
  email,
  password = "default",
  language = "en",
  divisionId = "DV_bf2LTv8JZn54Uh7",
  url = "QualtricsUrl"
)
}
\arguments{
\item{username}{User ID}

\item{firstName}{User first name}

\item{lastName}{User last name}

\item{userType}{The internal id of the user type. Default is id for "default access".}

\item{email}{User email}

\item{password}{User password}

\item{language}{User language}

\item{divisionId}{The internal id of the user's divison. Default is id for "Webster Worldwide".}

\item{url}{From Renviron OR alternative name in R.environ OR url of server}
}
\description{
Get user details for the provided user id.
}
