% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customSBCollapsePanel.R
\name{customSBCollapsePanel}
\alias{customSBCollapsePanel}
\title{Custom Shiny Collapse Panel}
\usage{
customSBCollapsePanel(name, content_tag_list)
}
\arguments{
\item{name}{A name for the panel, must be unique}

\item{content_tag_list}{Content stored in a shiny tagList}
}
\value{

}
\description{
Custom Shiny Collapse Panel
}
