% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfxns.R
\name{density_calculate}
\alias{density_calculate}
\title{Calculate density}
\usage{
density_calculate(
  coordinates,
  x = "UMAP_1",
  y = "UMAP_2",
  xlims = range(coordinates[[x]]),
  ylims = range(coordinates[[y]]),
  bins = 100,
  bw = "bcv",
  custom = NULL,
  rel_threshold = 0,
  abs_threshold = 0
)
}
\arguments{
\item{coordinates}{}

\item{x}{}

\item{y}{}

\item{xlims}{}

\item{ylims}{}

\item{bins}{}

\item{bw}{}

\item{custom}{}

\item{rel_threshold}{}

\item{abs_threshold}{}
}
\description{
Calculate density
}
