% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfxns.R
\name{density_plot}
\alias{density_plot}
\title{Plot normalized density}
\usage{
density_plot(
  input,
  x = "UMAP_1",
  y = "UMAP_2",
  xlims = NULL,
  ylims = NULL,
  group = NULL,
  target = NULL,
  background = "all",
  bins = 100,
  bw = "custom",
  custom = NULL,
  expansion = 0.05,
  plot = TRUE,
  rel_threshold = 0,
  abs_threshold = 0,
  downsample_data = 1,
  downsample_points = 1,
  point_alpha = 0.1,
  point_size = 0.1,
  point_color = "grey90",
  clip_zeros = TRUE,
  colors = "custom1",
  background_color = "transparent",
  borders = TRUE,
  table_only = FALSE
)
}
\arguments{
\item{input}{}

\item{x}{}

\item{y}{}

\item{xlims}{}

\item{ylims}{}

\item{group}{}

\item{target}{}

\item{background}{}

\item{bins}{}

\item{bw}{}

\item{custom}{}

\item{expansion}{}

\item{plot}{}

\item{rel_threshold}{}

\item{abs_threshold}{}

\item{downsample_points}{}

\item{point_alpha}{}

\item{point_size}{}

\item{point_color}{}

\item{clip_zeros}{}

\item{colors}{}

\item{background_color}{}

\item{borders}{}

\item{table_only}{}
}
\description{
Plot normalized density
}
