% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{enrichment_test}
\alias{enrichment_test}
\title{Enrichment testing}
\usage{
enrichment_test(
  list1,
  list2,
  cat1 = c(TRUE, FALSE),
  cat2 = c(TRUE, FALSE),
  background,
  print = TRUE,
  method = "fisher"
)
}
\arguments{
\item{list1}{}

\item{list2}{}

\item{cat1}{}

\item{cat2}{}

\item{background}{}

\item{print}{}
}
\description{
Enrichment testing
}
