% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfxns.R
\name{extract_expression}
\alias{extract_expression}
\title{Title}
\usage{
extract_expression(
  seuratobj,
  cells = NULL,
  genes,
  total = TRUE,
  normalize = TRUE,
  log = FALSE,
  assay = "RNA",
  slot = "counts"
)
}
\arguments{
\item{seuratobj}{}

\item{cells}{}

\item{genes}{}

\item{total}{}

\item{normalize}{}

\item{log}{}

\item{assay}{}

\item{slot}{}
}
\description{
Title
}
