% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQSurveys.R
\name{getQSurveys}
\alias{getQSurveys}
\title{Get all Qualtrics surveys}
\usage{
getQSurveys(url = "QualtricsUrl")
}
\arguments{
\item{url}{From Renviron OR alternative name in R.environ OR url of server}
}
\description{
Get surveys details for the token owner
}
