% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQUser.R
\name{getQUser}
\alias{getQUser}
\title{Get a specific Qualtrics user}
\usage{
getQUser(userID, url = "QualtricsUrl")
}
\arguments{
\item{userID}{User ID to get details for}

\item{url}{From Renviron OR alternative name in R.environ OR url of server}
}
\description{
Get user details for the provided user id.
}
