% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQUsers.R
\name{getQUsers}
\alias{getQUsers}
\title{Get all Qualtrics users}
\usage{
getQUsers(url = "QualtricsUrl")
}
\arguments{
\item{url}{From Renviron OR alternative name in R.environ OR url of server}
}
\description{
Get all user details
}
