% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSectionCourse.R
\name{getSectionCourse}
\alias{getSectionCourse}
\title{Check Section for Course}
\usage{
getSectionCourse(wcrid, server = "test", verbose = FALSE)
}
\arguments{
\item{wcrid}{Canvas course SIS ID}

\item{server}{Beta, test, or production}

\item{verbose}{Boolean to see API calls in the console}
}
\value{

}
\description{
Takes a Canvas SIS course ID and checks it as a section ID to determine what
course it is a part of.
}
