% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{ggsave2}
\alias{ggsave2}
\title{Title}
\usage{
ggsave2(
  plot,
  path = here("output/plots"),
  h = 4,
  w = 6,
  filename,
  dpi = 150,
  device = cairo_pdf,
  type = c("png", "pdf"),
  ...
)
}
\arguments{
\item{plot}{}

\item{path}{}

\item{h}{}

\item{w}{}

\item{filename}{}

\item{dpi}{}

\item{device}{}

\item{...}{}
}
\description{
Title
}
