% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQParamList.R
\name{makeQParamList}
\alias{makeQParamList}
\title{Create a Live Excel Parameter List}
\usage{
makeQParamList(parsedUrl)
}
\arguments{
\item{parsedUrl}{Parsed url from the parseIqyUrl function}
}
\description{
This function uses a parsed URL to identify the needed parameters for a query.  This is used
to supply the search parameters without being prompted.
}
\examples{
url <- parseIqyUrl("../path/to/query.iqy")
parameters <- makeQParamList(url)
parameters  ##To look at the requirements

parameters$parameter_0 <- "2018"
}
