% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{odds_quick}
\alias{odds_quick}
\title{Title}
\usage{
odds_quick(group1, group2, overlap, total, clean = TRUE)
}
\arguments{
\item{group1}{}

\item{group2}{}

\item{overlap}{}

\item{total}{}

\item{clean}{}
}
\description{
Title
}
