% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openIqy.R
\name{openIqy}
\alias{openIqy}
\title{Open an Informer Live Excel .iqy file}
\usage{
openIqy(path, param_list = NULL, build_only = FALSE)
}
\arguments{
\item{path}{Character of the path to a Live Excel .iqy file.}

\item{param_list}{A named list with the run time parameters for opening the data source.  If not provided,
these fields will be prompted for in the console.}

\item{build_only}{If TRUE, the function can be used as a helper app to build the param_list without executing the query.}
}
\description{
This function is designed to take the path to an .iqy file used with Excel.  The function can
be provided a named list of predifined parameters.
}
