% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCanvasUrl.R
\name{parseCanvasUrl}
\alias{parseCanvasUrl}
\title{Parse a Canvas URL}
\usage{
parseCanvasUrl(url)
}
\arguments{
\item{url}{}
}
\value{

}
\description{
Extract a named vector of IDs from a Canvas URL.
}
