% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseIqyUrl.R
\name{parseIqyUrl}
\alias{parseIqyUrl}
\title{Live Excel IQY Parser}
\usage{
parseIqyUrl(path)
}
\arguments{
\item{path}{Character of the path to an Informer generated "Live Excel" .iqy file.}
}
\description{
Live Excel IQY Parser
}
\examples{
url <- parseIqyUrl("../path/to/query.iqy")
}
