% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Title}
\usage{
plot_sankey(
  df,
  nodes,
  names = NULL,
  reverse = TRUE,
  total = TRUE,
  prefix = "Interactions:",
  arrange = 1,
  title = NULL,
  drop_na = FALSE
)
}
\arguments{
\item{df}{}

\item{nodes}{}

\item{names}{}

\item{reverse}{}

\item{total}{}

\item{prefix}{}

\item{arrange}{}

\item{title}{}

\item{drop_na}{}
}
\description{
Title
}
