% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{plot_sankey1}
\alias{plot_sankey1}
\title{Title}
\usage{
plot_sankey1(
  df,
  nodes,
  names = NULL,
  reverse = TRUE,
  top_n = 3,
  percentages = TRUE,
  total = TRUE,
  prefix = "Interactions:",
  title = NULL,
  drop_na = FALSE,
  ordered = FALSE
)
}
\arguments{
\item{df}{}

\item{nodes}{}

\item{names}{}

\item{reverse}{}

\item{top_n}{}

\item{percentages}{}

\item{total}{}

\item{prefix}{}

\item{title}{}
}
\description{
Title
}
