% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{plot_sankey2}
\alias{plot_sankey2}
\title{Title}
\usage{
plot_sankey2(
  df,
  nodes,
  names = NULL,
  top_n = 3,
  percentages = TRUE,
  total = TRUE,
  prefix = "Interactions:",
  title = NULL,
  drop_na = FALSE
)
}
\arguments{
\item{df}{}

\item{nodes}{}

\item{names}{}

\item{top_n}{}

\item{percentages}{}

\item{total}{}

\item{prefix}{}

\item{title}{}
}
\value{

}
\description{
Title
}
