% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfxns.R
\name{pseudobulk}
\alias{pseudobulk}
\title{Title}
\usage{
pseudobulk(
  seuratobj,
  genes,
  cells = NULL,
  group_by = "ident",
  pseudocount = 1,
  rounding = 3,
  assay = "RNA",
  slot = "count"
)
}
\arguments{
\item{seuratobj}{}

\item{genes}{}

\item{cells}{}

\item{group_by}{}

\item{pseudocount}{}

\item{rounding}{}

\item{assay}{}

\item{slot}{}
}
\description{
Title
}
