% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfxns.R
\name{pseudobulk_matrix}
\alias{pseudobulk_matrix}
\title{Title}
\usage{
pseudobulk_matrix(seuratobj, cells = NULL, group_by = "ident")
}
\arguments{
\item{seuratobj}{}

\item{cells}{}

\item{group_by}{}
}
\value{

}
\description{
Title
}
