% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfxns.R
\name{seurat_feature}
\alias{seurat_feature}
\title{Feature plot}
\usage{
seurat_feature(
  seuratobj,
  features = "ident",
  cells = NULL,
  facets = NULL,
  nrow = NULL,
  continuous = "auto",
  show = TRUE,
  alpha = "auto",
  size = "auto",
  reduction = "umap",
  dims = 1:2,
  title = NULL,
  title_params = "italics",
  color_package = c("custom", "generate", "carto", "ggplot"),
  color_palette = 1,
  borders = TRUE,
  na_break = TRUE,
  normalize = TRUE,
  scaled = FALSE,
  label = TRUE,
  label_color = c("white", "black"),
  label_size = 5,
  drop_points = FALSE,
  axis_title_position = 0,
  legend_position = "auto",
  legend_size = 10,
  facet_background = "dodgerblue4",
  facet_color = "white",
  facet_size = 10,
  facet_hide = FALSE,
  rasterize = "auto",
  rasterize_scale = 0.75,
  rasterize_dpi = 300,
  rasterize_threshold = 10000,
  verbose = TRUE
)
}
\arguments{
\item{seuratobj}{}

\item{features}{}

\item{cells}{}

\item{facets}{}

\item{nrow}{}

\item{continuous}{}

\item{show}{}

\item{alpha}{}

\item{size}{}

\item{reduction}{}

\item{dims}{}

\item{title}{}

\item{title_params}{}

\item{color_package}{}

\item{color_palette}{}

\item{borders}{}

\item{na_break}{}

\item{normalize}{}

\item{scaled}{}

\item{label}{}

\item{label_color}{}

\item{legend_position}{}

\item{legend_size}{}

\item{facet_background}{}

\item{facet_color}{}

\item{facet_size}{}

\item{verbose}{}
}
\description{
Feature plot
}
