% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfxns.R
\name{seurat_statsplot}
\alias{seurat_statsplot}
\title{Statsplot}
\usage{
seurat_statsplot(
  seuratobj,
  x = "nCount_RNA",
  y = "nFeature_RNA",
  marginal.type = "density",
  size = 0.5,
  alpha = 0.05,
  title = seuratobj@project.name,
  cells = NULL,
  logx = FALSE,
  logy = FALSE,
  vline = NULL,
  hline = NULL,
  xlab = NULL,
  ylab = NULL,
  median_lines = FALSE,
  xlim = "auto",
  ylim = "auto",
  ...
)
}
\arguments{
\item{seuratobj}{}

\item{x}{}

\item{y}{}

\item{marginal.type}{}

\item{size}{}

\item{alpha}{}

\item{title}{}

\item{cells}{}

\item{logx}{}

\item{logy}{}

\item{vline}{}

\item{hline}{}

\item{median_lines}{}

\item{xlim}{}

\item{ylim}{}

\item{...}{}
}
\description{
Statsplot
}
