% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabTitleBox.R
\name{tabTitleBox}
\alias{tabTitleBox}
\title{Tab Title Box}
\usage{
tabTitleBox(title, details, modal = TRUE)
}
\arguments{
\item{title}{The title of the tab.}

\item{details}{Test to appear that describes the purpose of the tab.}

\item{modal}{Boolean optional "more info" modal}
}
\value{

}
\description{
Add extra functionality to titles, including a "more info" modal. Provides a
title and text to appear in a well under it.  A more info button
added as an observe.
}
