% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_dwu}
\alias{theme_dwu}
\title{Custom theme}
\usage{
theme_dwu(
  base_size = 12,
  base_family = "Arial",
  border = TRUE,
  axis = TRUE,
  grid = FALSE,
  legend.position = "none",
  facet_background = "grey80",
  facet_border = "black",
  facet_color = "black"
)
}
\arguments{
\item{base_size}{}

\item{base_family}{}

\item{border}{}

\item{axis}{}

\item{grid}{}

\item{legend.position}{}

\item{facet_background}{}

\item{facet_border}{}

\item{facet_color}{}
}
\description{
Custom theme
}
