% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateQUser.R
\name{updateQUser}
\alias{updateQUser}
\title{Update a specific Qualtrics user}
\usage{
updateQUser(
  userID,
  userType = "Standard - Webster University",
  accountExpirationDate = NULL,
  url = "QualtricsUrl"
)
}
\arguments{
\item{userID}{User ID to update details for}

\item{userType}{The type of user which determines privileges.  "Default access" is the default state for students.  "Standard - Webster University" is the default for faculty/staff.}

\item{accountExpirationDate}{Set the date that the access expires (yyyy-mm-dd)}

\item{url}{From Renviron OR alternative name in R.environ OR url of server}
}
\description{
Get user details for the provided user id.
}
