% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_publication}
\alias{theme_publication}
\title{Theme for publication}
\usage{
theme_publication(
  base_size = 12,
  base_family = "Arial",
  axis = T,
  grid = F,
  legend.position = "none"
)
}
\arguments{
\item{base_size}{}

\item{base_family}{}

\item{axis}{}

\item{grid}{}

\item{legend}{}
}
\description{
Theme for publication
}
